server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bcrypt<4",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "mlx-vlm>=0.4.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]
